rename: profile 통합
